BetterPosts: Al enviar un post, borrar la copia solo si se tiene activa la preferencia de guardar copias.
Si no se tiene activada la opción de guardar los posts no enviados no
tiene sentido eliminar del servidor los datos guardados (ya que se
supone que no habrá nada).

+++ modules/BetterPosts.js
@@ -441,10 +441,12 @@
 
 					if ((timeToWait--) <= 0) {
 						clearInterval(interval);
-						mod.helper.deleteValue("POST_BACKUP");
-						$("form[name='vbform']").submit();
+						if (mod.preferences.savePosts) {
+							mod.helper.deleteValue("POST_BACKUP");
+						}
+						document.getElementsByName('vbform')[0].submit();
 					} else {
-						errors.html("Debes esperar al menos 30 segundos entre cada envio de nuevos mensajes. El mensaje se enviará automáticamente en " + (timeToWait) + " segundos. <a style='color: #CC3300;cursor:pointer;' onclick='clearInterval(autoReplyInterval); this.remove();'>Cancelar</a>");
+						errors.html("Debes esperar al menos 30 segundos entre cada envio de nuevos mensajes. El mensaje se enviará automáticamente en " + timeToWait + " segundos. <a style='color: #CC3300;cursor:pointer;' onclick='clearInterval(autoReplyInterval); this.remove();'>Cancelar</a>");
 					}
 
 				} else {
